requirements.txt
aiohttp==3.8.4
pydantic==1.10.7
python-telegram-bot==20.2
SQLAlchemy==2.0.7
